frontend/requirements.txt
Flask==2.0.1
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf>=4.21.0
